178; deleted MatchManager as it was redundant; copied shared_utilities/GameSetUpManager contents into Match/managers.py; deleted GameSetUpmanager file
