+++ backend/Match/managers.py
@@ -1,28 +1,42 @@
 from django.db import models, transaction, IntegrityError
-from django.db.models import Q
 
-from .playerMatchStatus import PlayerMatchStatus
-from Player.models import Player
+from .models import Match
+from Tokens.models import MatchToken
 from User.models import User
+from Player.models import Player
 
 import sys
 
-class MatchManager(models.Manager):
-	def get_matches_with_user(self, user_id, match_status):
+class GameSetupManager:
+	@staticmethod
+	def create_match_and_its_players(match_token):
+		if not isinstance(match_token, MatchToken):
+			raise TypeError((f'match_token must be an MatchToken, not {type(match_token).__name__}'))
+		
 		try:
-			user = User.objects.get(pk=user_id)
-		except User.DoesNotExist:
-			raise ValueError("User with id '{}' does not exist.".format(user_id))
-
-		matches_with_user = self.filter(players__user_id=user_id)
-
-		if match_status == PlayerMatchStatus.WINS.value:
-			matches_with_user = matches_with_user.filter(
-				players__user_id=user_id, players__match_winner=True
-			)
-		elif match_status == PlayerMatchStatus.LOSSES.value:
-			matches_with_user = matches_with_user.filter(
-				players__user_id=user_id, players__match_winner=False
-			)
-
-		return matches_with_user
+			with transaction.atomic():
+				# modify later to take into consideration tournament data
+				match = Match.objects.create()
+				
+				player_left_side = Player.objects.create(
+					user_id=match_token.user_left_side.id,
+					match=match,
+					score=0,
+					match_winner=False
+				)
+				
+				player_right_side = Player.objects.create(
+					user_id=match_token.user_right_side.id,
+					match=match,
+					score=0,
+					match_winner=False
+				)
+		
+		except IntegrityError as e:
+			print(f"integrity error: An error occurred: {e}", file=sys.stderr)
+			return None, None, None
+		except Exception as e:
+			print(f"An error occurred: {e}", file=sys.stderr)
+			return None, None, None
+		
+		return match, player_left_side, player_right_side
