+++ backend/Match/models.py
@@ -19,8 +19,6 @@ class Match(models.Model):
 	created_at = models.DateTimeField(auto_now_add=True)
 	updated_at = models.DateTimeField(auto_now=True)
 
-	objects = MatchManager()
-
 	def start_match(self):
 		if self.state == MatchState.LOBBY.value:
 			self.state = MatchState.IN_PROGRESS.value
